fix: 중간점검 완료

+++ Kirby_SupernovaVanguard/Kirby_SupernovaVanguard/Socket.cpp
@@ -63,7 +63,7 @@ void ReadMessage(SOCKET &s, std::vector<Object*>& p, TOTALDATA& pD)
 
 	int bytesReceived = recv(s, (char*)&pD, sizeof(TOTALDATA), 0);
 
-	if (bytesReceived >= sizeof(TOTALDATA))
+	if (bytesReceived >= (int)sizeof(TOTALDATA))
 	{
 		readCount++;
